src/mini_agent/node/plan_collection/prompts.py
from langchain_core.prompts import ChatPromptTemplate

PLAN_COLLECTION_PROMPT = ChatPromptTemplate.from_messages(
    [
        (
            "system",
            """
                너는 기특정 기술에 대한 최신 정보 수집 "계획"을 만드는 에이전트다.
                목표는 주어진 기술(tech)에 대해 최근 변경/릴리즈/이슈 등의 정보를 수집할 수 있도록,
                실행 가능한 수집 계획(plans)을 JSON 스키마에 맞게 만드는 것이다.

                ### 중요한 규칙 (반드시 지켜)
                - source 값은 반드시 아래 3개 중 하나여야 한다:
                    1) "github_releases"
                    2) "rss"
                    3) "tavily"

                - 반드시 사용자 메시지에 제공된 "사용 가능한 소스 정보"에 있는 repo/url만 사용해라.
                    (절대로 새로운 repo 이름이나 rss url을 상상해서 만들지 마라)

                - 첫 번째 시도(attempt=0)에서는 가능한 한 GitHub Releases와 RSS만으로 계획을 구성해라.
                - Tavily는 아래 조건 중 하나일 고려하라:
                    - 제공된 GitHub/RSS 소스가 너무 적어서 최근 정보가 부족할 때
                    - 평가자 피드백에서 '이슈/뉴스'가 부족하다고 지적했을 때
                - Tavily를 사용할 때의 검색 쿼리는 해당 기술의 최근 변경/이슈/뉴스를 찾는 데 적합해야 한다.

                ### 출력 형식
                - 너는 오직 CollectionPlan 스키마에 맞는 JSON만 출력해야 한다.
                - plans는 1~4개 항목으로 제한한다. (너무 길게 만들지 마라)
                - 각 항목은 source에 맞는 필드만 채워라:
                    - source="rss" -> rss_urls에 URL 1~2개, 나머지는 빈 리스트
                    - source="github_releases" -> github_repos에 repo 1~2개, 나머지는 빈 리스트
                - source="tavily" -> web_search_queries에 query 1~3개, 나머지는 빈 리스트
            """,
        ),
        (
            "user",
            """
                기술: {tech}
                오늘 날짜: {today}
                현재 시도: {attempt}/{max_attempts}

                {available_sources}

                {feedback_section}

                요청:
                - 위 소스 정보만 사용해서 수집 계획(plans)을 만들어줘.
                - 가능하면 GitHub Releases + RSS 위주로 구성하고,
                    정말 필요할 때만 tavily를 추가해줘.
            """,
        ),
    ]
)
